added:UserId

+++ src/services/note.service.js
@@ -2,14 +2,14 @@ import { error } from '@hapi/joi/lib/base';
 import Notes from '../models/note.model';
 import { userAuth } from '../middlewares/auth.middleware';
 
-export const newNote = async (body) => {
-    const data = await Notes.create(body);
+export const newNote = async (body,userId) => {
+    const data = await Notes.create({...body,userId:userId});
     // console.log(data)
     return data;
 }
 
-export const updateNote = async (body, id) => {
-    let data = await Notes.findOneAndUpdate({ _id: id }, body, { new: true })
+export const updateNote = async (body, id,userId) => {
+    let data = await Notes.findOneAndUpdate({ _id: id,userId,userId }, body, { new: true })
     if (!data) {
         throw new error("Not found")
     }
@@ -19,24 +19,24 @@ export const updateNote = async (body, id) => {
 
 //Get all the notes
 export const getAll = async (userId) => {
-    console.log(userId);
-    const note = await Notes.find(userId)
+    
+    const note = await Notes.find({userId:userId})
     return note;
 }
 
 //delete the Onenote
 export const delNote = async (id,userId) => {
-    const data = await Notes.findByIdAndDelete(id,userId)
+    const data = await Notes.findByIdAndDelete({_id:id,userId:userId})
     return data;
 }
 
 //ArchiveNote
 export const archiveNote = async (noteId) => {
-    const existNoteId = await Notes.findOne({ _id: noteId },userId)
+    const existNoteId = await Notes.findOne({ _id: noteId },{userId:userId})
     if (!existNoteId) {
         throw new Error('Note Not Found');
     }
-    const updateNote = await Notes.findOneAndUpdate((noteId),userAuth.id,
+    const updateNote = await Notes.findOneAndUpdate((noteId),{userId:userId},
         { $set: { archive: !existNoteId.archive } },
         { new: true })
 
@@ -44,11 +44,11 @@ export const archiveNote = async (noteId) => {
 }
 //trashNote
 export const trashNote = async (noteId) => {
-    const existNoteId = await Notes.findOne({ _id: noteId },userAuth.id)
+    const existNoteId = await Notes.findOne({ _id: noteId },{userId:userId})
     if (!existNoteId) {
         throw new Error('Note Not Found');
     }
-    const updateNote = await Notes.findOneAndUpdate((noteId),userAuth.id,
+    const updateNote = await Notes.findOneAndUpdate((noteId),{userId:userId},
         { $set: { trash: !existNoteId.trash } }
         , { new: true })
 
@@ -56,7 +56,7 @@ export const trashNote = async (noteId) => {
 }
 
 //get single Note
-export const getNote = async (id) => {
-    const data = await Notes.findById(id,userAuth.id);
+export const getNote = async (_id,userId) => {
+    const data = await Notes.findOne({_id:_id, userId:userId});
     return data;
   };
